feat: More 버튼 수정
